hw8/requirements_server.txt
google-cloud-storage>=2.10.0
google-cloud-logging>=3.0.0
google-cloud-pubsub>=2.18.0
